Extract polling logic to checkoutPolling.ts to fix bun test suite resolution
